backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
openai==1.3.5
python-dotenv==1.0.0
trimesh==4.0.0
numpy==1.26.2
